apps/core/users/serializers.py
from rest_framework import serializers

from .models import User

class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True)

    def create(self, validated_data):
        user = User.objects.create_user(
            email=validated_data['email'],
            first_name=validated_data['first_name'],
            password=validated_data['password'],
        )
        return user

    class Meta:
        model = User
        fields = ( "id", "email", "first_name", "password", )