requirements.txt
anyio==3.7.1
APScheduler==3.6.3
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.3
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pycparser==2.21
PyJWT==2.8.0
python-telegram-bot==13.1
pytz==2022.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.7.1
tzlocal==5.0.1
